dashboard/requirements.txt
dash
pandas
sqlalchemy
psycopg2-binary
PyYAML